Adição e ajuste de ferramentas
Adição do BigDecimal ativamente para cálculos exatos, sem depender do double que pode gerar valores indesejados. Inserção do ValorInvalidoException no public ContaBancaria e no depositar. Transacao agora recebe BigDecimal valor para permitir as devidas operações.

+++ src/Transacao.java
@@ -1,4 +1,5 @@
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record Transacao(TipoTransacao tipoTransacao, double valor, LocalDateTime dataHora) {}
+public record Transacao(TipoTransacao tipoTransacao, BigDecimal valor, LocalDateTime dataHora) {}
